Removed all services from provide-viewer-instance. State as provideState
