+++ projects/core/src/lib/core.module.ts
@@ -1,7 +1,7 @@
 import {  ModuleWithProviders, NgModule, inject } from '@angular/core';
 import { PasswordResetComponent, LoginComponent, ViewerAppComponent } from './pages';
 import { MapModule } from '@tailormap-viewer/map';
-import { StoreModule } from '@ngrx/store';
+import { provideState, StoreModule } from '@ngrx/store';
 import { coreReducer } from './state/core.reducer';
 import { coreStateKey } from './state/core.state';
 import {
@@ -33,7 +33,8 @@ import { CoreRoutingModule } from './core-routing.module';
 import { AuthenticatedUserService } from '@tailormap-viewer/api';
 import { UserLoginCheckService } from './services/user-login-check.service';
 import { CoreSharedModule } from './shared/core-shared.module';
-import { getViewerFeatureStateProviders } from './viewer-instance/provide-viewer-instance';
+import { mapReducer } from './map/state/map.reducer';
+import { mapStateKey } from './map/state/map.state';
 
 const getBaseHref = (platformLocation: PlatformLocation): string => {
   return platformLocation.getBaseHrefFromDOM();
@@ -75,7 +76,7 @@ const getBaseHref = (platformLocation: PlatformLocation): string => {
     RouterModule,
   ],
   providers: [
-    ...getViewerFeatureStateProviders(),
+    provideState(mapStateKey, mapReducer),
     { provide: HTTP_INTERCEPTORS, useClass: SecurityInterceptor, multi: true },
     { provide: TAILORMAP_SECURITY_API_V1_SERVICE, useClass: TailormapSecurityApiV1Service },
     { provide: TAILORMAP_API_V1_SERVICE, useClass: TailormapApiV1Service },
